[PR] frontend 브랜치 데이터 추가

+++ Backend/erpc/src/main/java/com/cineverse/erpc/salesopp/note/dto/SalesOppNoteDTO.java
@@ -10,7 +10,7 @@ import lombok.*;
 @ToString
 @Builder
 public class SalesOppNoteDTO {
-    private int salesOppNoteId;
+    private long salesOppNoteId;
     private String salesOppNote;
     private String salesOppNoteDate;
     private String salesOppNoteDeleteDate;
